# apps/platform-api/specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API Documentation",
    "version": "1.0.0",
    "description": "API Documentation"
  },
  "paths": {
...
    "/api/v1/user/heartbeat": {
      "put": {
        "summary": "发送心跳，更新在线状态",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/user/me": {
      "get": {
        "summary": "获取当前用户完整资料",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "当前用户完整资料",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "object"
                    },
                    "preferredLocale": {
                      "type": "string",
                      "enum": [
                        "zh-CN",
                        "en"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "NORMAL",
                        "MUTED",
                        "BANNED",
                        "DELETED"
                      ]
                    },
                    "tokenVersion": {
                      "type": "number"
                    },
                    "lastLoginAt": {
                      "type": "object"
                    },
                    "nickname": {
                      "type": "object"
                    },
                    "avatarUrl": {
                      "type": "object"
                    },
                    "gender": {
                      "type": "object"
                    },
                    "birthday": {
                      "type": "object"
                    },
                    "bio": {
                      "type": "object"
                    },
                    "website": {
                      "type": "object"
                    },
                    "location": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "email",
                    "preferredLocale",
                    "status",
                    "tokenVersion",
                    "lastLoginAt",
                    "nickname",
                    "avatarUrl",
                    "gender",
                    "birthday",
                    "bio",
                    "website",
                    "location"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/me/profile": {
      "put": {
        "summary": "更新当前用户资料",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "avatar_url": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "website": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "Male",
                      "Female",
                      "Unknown"
                    ]
                  },
                  "birthday": {
                    "type": "string"
                  },
                  "preferred_locale": {
                    "type": "string",
                    "enum": [
                      "zh-CN",
                      "en"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "资料更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "affectedRows": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "affectedRows"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/me/email/send-code": {
      "post": {
        "summary": "发送邮箱验证码（更换邮箱）",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验证码已发送"
          }
        }
      }
    },
    "/api/v1/user/me/email": {
      "put": {
        "summary": "验证并更新邮箱",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 6
                  }
                },
                "required": [
                  "email",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "邮箱更新成功"
          }
        }
      }
    },
    "/api/v1/user/public/{username}": {
      "get": {
        "summary": "获取用户公开资料",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户公开资料",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "githubConnected": {
                      "type": "boolean"
                    },
                    "preferredLocale": {
                      "type": "string",
                      "enum": [
                        "zh-CN",
                        "en"
                      ]
                    },
                    "nickname": {
                      "type": "object"
                    },
                    "avatarUrl": {
                      "type": "object"
                    },
                    "gender": {
                      "type": "object"
                    },
                    "birthday": {
                      "type": "object"
                    },
                    "bio": {
                      "type": "object"
                    },
                    "website": {
                      "type": "object"
                    },
                    "location": {
                      "type": "object"
                    },
                    "updatedAt": {
                      "type": "object"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "githubConnected",
                    "preferredLocale",
                    "nickname",
                    "avatarUrl",
                    "gender",
                    "birthday",
                    "bio",
                    "website",
                    "location",
                    "updatedAt",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/{id}/brief": {
      "get": {
        "summary": "获取用户简档（供评论作者卡片使用）",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户简档信息"
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}